Split async process preamble helpers
