+++ crates/sifr_codegen/src/preamble.rs
@@ -14,8 +14,10 @@ mod join_set_runtime;
 pub use join_set_runtime::*;
 mod parallel_runtime;
 pub(crate) use parallel_runtime::*;
+mod process_async_runtime;
 mod process_child_pipes;
 mod process_runtime;
+pub(crate) use process_async_runtime::*;
 pub(crate) use process_runtime::*;
 mod io_logging_random;
 pub use io_logging_random::*;
